Moved show-ui modules to show_mode package
