+++ src/view/show_mode/_scene_selector_widget.py
