+++ src/view/show_mode/_ui_area_widget.py
