+++ src/view/show_mode/nodeeditor.py
@@ -1,0 +1,6 @@
+
+# coding=utf-8
+"""Module containing node editor"""
+class NodeEditorWidget:
+    """Nodeeditor to edit scenes and their filter nodes"""
+    pass
