+++ src/view/show_mode/show_ui_player_widget.py
